Move rest_api to indra_db_service.
