+++ indra_db_service/data-vis/src/components/AmountView/LineChart.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-  import VueApexCharts from 'vue-apexcharts'
+  import VueApexCharts from 'vue3-apexcharts';
 
   export default {
     name: "LineChart",
