+++ indra_db_service/data-vis/src/components/AmountView/index.js
