+++ indra_db_service/data-vis/src/index.js
@@ -1,0 +1,2 @@
+export {default as TimeView} from './components/TimeView'
+export {default as AmountView} from './components/AmountView'
