+++ indra_db_service/data-vis/src/main.js
@@ -1,0 +1,8 @@
+import { createApp } from 'vue'
+import App from './App.vue'
+import {TimeView, AmountView} from './index'
+
+const app = createApp(App)
+app.use(TimeView)
+app.use(AmountView)
+app.mount('#app')
