+++ indra_db_service/data-vis/vue.config.js
