+++ indra_db_service/static/curationFunctions.js
