+++ indra_db_service/templates/idbr_footer.html
