--- rest_api/data-vis/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "data-vis",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build --target lib --name DataVis src/index.js",
-    "watch": "npm run build -- --watch",
-    "test": "npm run watch -- --mode development",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "apexcharts": "^3.15.4",
-    "core-js": "^3.4.4",
-    "vue": "^2.6.10",
-    "vue-apexcharts": "^1.5.2",
-    "vue-multiselect": "^2.1.6"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.0",
-    "@vue/cli-plugin-eslint": "^4.1.0",
-    "@vue/cli-service": "^4.1.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "vue-template-compiler": "^2.6.10"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "rules": {},
-    "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ]
-}
