--- rest_api/data-vis/src/index.js
@@ -1,17 +1,0 @@
-import * as datavisComps from './components'
-
-const install = (Vue) => {
-  Object.values(datavisComps).forEach(comp => {
-    window.console.log('DataVis installing ' + comp.name);
-    Vue.use(comp);
-  });
-};
-
-if (typeof window != 'undefined' && window.Vue) {
-  install(window.Vue)
-}
-
-export default install
-
-export {default as TimeView} from './components/TimeView'
-export {default as AmountView} from './components/AmountView'
